requirements.txt
Django==1.8.3
django-filter==0.10.0
django-pgjson==0.3.1
django-pgjsonb==0.0.16
djangorestframework==3.2.1
Markdown==2.6.2
psycopg2==2.6.1
wheel==0.24.0
